app_backup/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
opencv-python-headless==4.8.0.76
numpy==1.26.4
python-dotenv==1.0.0
ultralytics
torch==2.2.0
tflite-runtime
tensorflow==2.14.0